requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
joblib
mysql-connector-python
numpy
